Added comments to assembly language + improved docs
The docs/design.md file now has documentation on some less obvious
aspects of the assembly language. This can be found in the "Assembly
syntax" section towards the end of the file.

Comments support was added to the assembly language. This meant that a
pre-processor, which runs before the lexer, had to be added.
Documentation for these can be found in the "Comments" section towards
the end of docs/design.md.