No recursive instruction execution via separate threads
The emulator now has a event thread or EmulatorThread which waits for
events, and handles them appropriately.
It also has a execution thread which will just run through the
instructions. It is killed and restarted when an instruction needs to be
interrupted such as when an interrupt (which could be the result of an
exception) occurs.

All of this eliminates recursion in the instruction execution code.