LibArch + complex memory addresses.
There is now a library for encoding and decoding assembly for this
architecture. It is quite simple, but the Assembler and Emulator use it.
The Emulator still does some stuff on it's own, that should be done by
the library.

The old register offset memory addresses have now been replaced by
complex memory addresses. This makes creating much more complex programs
possible.