General clean-up, BIOS region, assembly improvements
A large clean-up of the Emulator was done as there was a lot of old
useless code.

There is now a BIOS memory region starting at 0xF000'0000. This is the
new default entry point, and the program parsed as an argument is loaded
there. The region goes for 0x1000'0000 bytes, which is also the limit
for the program.

The assembler now has integer literals of bases other than 10 support.
It supports hex literals starting with 0x or 0X, binary literals
starting with 0b or 0B, and octal literals starting with 0O or 0o.