Video Device
There is now a video device which currently uses SDL2 as a backend. This
is a basic device, and not fully featured yet. Just has basic init, get
info, get mode, and set mode commands.

It comes with a I/O device and a memory region which can override a RAM
region for framebuffer usage.

Some bug fixes in the assembler and emulator were also made.