Hardware Interrupts
There is now support for hardware (device) interrupts. Devices can specify a
number of interrupts to the I/O bus, which handles routing them to
software (CPU) interrupts. This came with 2 new I/O bus commands and an
update to an existing one.

Also with this there was a fairly important bug fix relating to
termination of the execution thread. Instead of forcefully killing it,
it is now requested for it to die.

Another thing is that there is now a `hash_combine` function in new math
header/source files. This was taken from boost.