Minor bug fixes in the Preprocessor and Emulator
The bug in the preprocessor related to sorting of the define references
before replacing them. It now uses a std::map instead of a std:vector of
a std::pair so that it is already sorted.

The bug in the emulator was that a seg fault would occur if
Emulator::Crash was called before the registers were created. There is
now a check to prevent this.