Emulator performance improvements + minor assembler bug fix
Both LibArch and the emulator were modified to make the emulator run
faster. Most of this change was focused around less heap allocation.

A minor bug fix was made in the assembler's lexer. This was caused by
some code the was thought to be unnecessary getting removed in the
previous commit.