Better error messages in the Assembler + encode side of LibArch
A new error message system has been setup in all aspects of the
assembler. The file name and line number for different parts of the
source is stored in different ways across the different stages to allow
for much better error messages.

Various bugs and code style issues were fixed throughout the Assembler
and LibArch as well.